requirements.txt
requests==2.31.0
google-cloud-storage==2.14.0
pandas==2.2.0
pyarrow==15.0.0
polars==0.20.10
gcsfs==2023.10.0
fsspec==2023.10.0
plotly==5.18.0